refactor pt1
