+++ ui/litellm-dashboard/src/components/AIHub/mcp_hub_table_columns.tsx
